Complete Rock Paper Scissors Project
I finished the project. I will hold on to this version of the project
until The Odin Project calls for me to go back to it. Instead of
obsessing over small details and getting lost in the weeds, I think it
would be better to keep moving forward and learning new things.